% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim.R
\name{fim}
\alias{fim}
\title{fim
Return the specified Fisher Information Matrix, given design matrix, coefficient vector, and model choice}
\usage{
fim(X, beta, model)
}
\arguments{
\item{X}{matrix}

\item{beta}{numeric}

\item{model}{character}
}
\value{
matrix
}
\description{
fim
Return the specified Fisher Information Matrix, given design matrix, coefficient vector, and model choice
}
