% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_weights_logit.R
\name{fim_weights_logit}
\alias{fim_weights_logit}
\title{fim_weights_logit
Computes the Fisher Information Matrix weights for a logistic regression model}
\usage{
fim_weights_logit(X, beta)
}
\arguments{
\item{X}{matrix The design matrix of the model}

\item{beta}{vector The coefficient vector}
}
\value{
vector
}
\description{
fim_weights_logit
Computes the Fisher Information Matrix weights for a logistic regression model
}
