% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mde.R
\name{mde}
\alias{mde}
\title{mde
Compute the minimum detectable effect for a logit model}
\usage{
mde(power, n, X, beta, j, upper = 100, sig = 0.05)
}
\arguments{
\item{power}{numeric Desired power level}

\item{n}{numeric Sample size}

\item{X}{matrix Design matrix}

\item{beta}{numeric Coefficient vector}

\item{j}{numeric Coefficient to estimate MDE for}

\item{upper}{numeric Upper bound on MDE search, default is 100 TODO: choose this smarter or eliminate}

\item{sig}{numeric Statistical power, default is .05}
}
\value{
numeric
}
\description{
mde
Compute the minimum detectable effect for a logit model
}
