% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mde.R
\name{mde_objective}
\alias{mde_objective}
\title{mde_objective
An objective function passed to the root finder which solves for the MDE}
\usage{
mde_objective(MDE, power, n, X, beta, j, sig = 0.05)
}
\arguments{
\item{MDE}{numeric main argument of the objective}

\item{power}{numeric}

\item{n}{numeric}

\item{X}{matrix}

\item{beta}{numeric}

\item{j}{numeric}

\item{sig}{numeric}
}
\value{
numeric
}
\description{
mde_objective
An objective function passed to the root finder which solves for the MDE
}
