% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_power.R
\name{wald_power}
\alias{wald_power}
\title{wald.power
Compute the power of a Wald test statistic for the jth coefficient of parameter vector beta}
\usage{
wald_power(X, beta, j, sig = 0.05, model = "logit")
}
\arguments{
\item{X}{matrix The design matrix of the experiment}

\item{beta}{numeric The coefficient vector}

\item{j}{numeric The index of the coefficient whose power is to be estimated}

\item{sig}{numeric The statistical significance of the test, default is .05}

\item{model}{character The model type, default is 'logit'}
}
\value{
numeric
}
\description{
wald.power
Compute the power of a Wald test statistic for the jth coefficient of parameter vector beta
}
