% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{z.alpha}
\alias{z.alpha}
\title{Evaluate the 1-alpha/2 quantile of a std normal}
\usage{
z.alpha(alpha)
}
\arguments{
\item{alpha}{numeric}
}
\value{
numeric
}
\description{
Evaluate the 1-alpha/2 quantile of a std normal
}
