% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_logit}
\alias{inv_logit}
\title{Evaluate the inverse logit function, ie. the logitistic function}
\usage{
inv_logit(z)
}
\arguments{
\item{z}{numeric}
}
\value{
numeric
}
\description{
Evaluate the inverse logit function, ie. the logitistic function
}
